pyproject.toml
[project]
name = "code-debug-env"
version = "1.0.0"
description = "Python code debugging environment for RL agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.2.3",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "openai>=1.0.0",
]
